requirements.txt
chromadb==0.4.24
keras==3.2.1
nltk==3.8.1
pdfminer.rtl==1.0.1
spacy==3.7.4
tensorflow==2.16.1
transformers==4.40.0
